상품 페이지 경로변수 수정
쿼리파라미터를 사용 X, 대신 경로 변수를 사용

+++ src/main/java/com/appleshopingmall/shop/product/ProductEntity.java
@@ -6,7 +6,7 @@ import java.sql.Timestamp;
 
 @Data
 public class ProductEntity {
-    private long productID;
+    private int productID;
     private String productName;
     private long productPrice;
     private Timestamp productData;
